requirements.txt
streamlit
google-generativeai
pytest
playwright
python-dotenv
